requirements.txt
Pillow>=10.0.0
geopy>=2.4.0
